Moved get_script_folder to utils.io.
